requirements.txt
dj-database-url==0.5.0
django-extensions==3.1.1
django-localflavor==3.0.1
django-widget-tweaks==1.4.8
Django==2.2.19
djangorestframework==3.12.2
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
